fix(pci-billing): refactor columns
ref: DTCORE-2788
