+++ packages/manager/apps/pci-billing/src/components/consumption/InstanceList.component.tsx
@@ -6,53 +6,24 @@ import {
   useImages,
   useInstances,
 } from '@ovh-ux/manager-pci-common';
-import {
-  Datagrid,
-  DataGridTextCell,
-  DateFormat,
-  useDataGrid,
-  useFormattedDate,
-} from '@ovh-ux/manager-react-components';
+import { Datagrid, useDataGrid } from '@ovh-ux/manager-react-components';
 import { ShellContext } from '@ovh-ux/manager-react-shell-client';
 import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 import {
-  ODS_BUTTON_SIZE,
-  ODS_BUTTON_TYPE,
-  ODS_BUTTON_VARIANT,
   ODS_SPINNER_SIZE,
   ODS_TEXT_LEVEL,
   ODS_TEXT_SIZE,
 } from '@ovhcloud/ods-components';
-import {
-  OsdsButton,
-  OsdsPopover,
-  OsdsPopoverContent,
-  OsdsSpinner,
-  OsdsText,
-} from '@ovhcloud/ods-components/react';
+import { OsdsSpinner, OsdsText } from '@ovhcloud/ods-components/react';
 import { useContext, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { TInstance } from '@/api/hook/useConsumption';
 import { paginateResults } from '@/api/data/consumption';
-import InstanceDetailPopover from './InstanceDetailPopover.component';
-
-export type TInstanceConsumptionDetail = {
-  instanceId: string;
-  instanceName: string;
-  total: string;
-  region: string;
-  reference: string;
-  imageType: string;
-  vmType: string;
-  isDeleted: boolean;
-  monthlyBilling?: {
-    since: string;
-    status: string;
-  };
-  planCode?: string;
-  monthlyPlan: string[];
-};
+import {
+  TInstanceConsumptionDetail,
+  useInstanceListColumns,
+} from './useInstanceListColumns';
 
 type InstanceListProps = {
   billingInstances: TInstance[];
@@ -73,8 +44,13 @@ export default function InstanceList({
 
   const { currency } = useContext(ShellContext).environment.getUser();
   const { projectId } = useParams();
-  const navigate = useNavigate();
   const { pagination, setPagination } = useDataGrid();
+  const columns = useInstanceListColumns({
+    colNameLabel,
+    colTotalLabel,
+    showAdditionalInstanceDetails,
+    isMonthlyInstances,
+  });
 
   const [instanceConsumptionDetails, setInstanceConsumptionDetails] = useState({
     rows: [],
@@ -156,15 +132,6 @@ export default function InstanceList({
     return instanceConsumptionDetail;
   };
 
-  // TO DISPLAY CONFIRM MONTHLY PAYMENT BUTTON
-  const displayUpdateToMonthlyBillingButton = (
-    instanceConsumption: TInstanceConsumptionDetail,
-  ) =>
-    !isMonthlyInstances &&
-    instanceConsumption.monthlyBilling === null &&
-    !instanceConsumption.isDeleted &&
-    instanceConsumption.monthlyPlan;
-
   useEffect(() => {
     if (billingInstances && allInstances && allImages && projectCatalog) {
       const result = billingInstances.map((billingDetail) =>
@@ -192,72 +159,6 @@ export default function InstanceList({
     pagination,
     setPagination,
   ]);
-
-  const columns = [
-    {
-      id: 'name',
-      cell: (row: TInstanceConsumptionDetail) => (
-        <OsdsPopover>
-          <div slot="popover-trigger" className="cursor-pointer">
-            {row?.instanceName}
-          </div>
-          <OsdsPopoverContent>
-            <InstanceDetailPopover row={row} />
-          </OsdsPopoverContent>
-        </OsdsPopover>
-      ),
-      label: colNameLabel,
-    },
-    {
-      id: 'consumption',
-      cell: (row: TInstanceConsumptionDetail) => (
-        <div>
-          <DataGridTextCell>{row.total}</DataGridTextCell>
-        </div>
-      ),
-      label: colTotalLabel,
-    },
-    (!isMonthlyInstances || showAdditionalInstanceDetails) && {
-      id: 'action',
-      cell: (row: TInstanceConsumptionDetail) => (
-        <div className="min-w-[16rem]">
-          {['ok', 'resized'].includes(row.monthlyBilling?.status) && (
-            <DataGridTextCell>
-              {t('cpbc_hourly_instance_monthly_billing_since', {
-                since: useFormattedDate({
-                  dateString: row.monthlyBilling.since,
-                  format: DateFormat.compact,
-                }),
-              })}
-            </DataGridTextCell>
-          )}
-          {row.monthlyBilling?.status === 'activationPending' && (
-            <DataGridTextCell>
-              {t('cpbc_hourly_instance_monthly_billing_activationPending')}
-            </DataGridTextCell>
-          )}
-          {!row.monthlyBilling && row.isDeleted && (
-            <DataGridTextCell className="opacity-50">
-              {t('cpbc_hourly_instance_deleted')}
-            </DataGridTextCell>
-          )}
-
-          {displayUpdateToMonthlyBillingButton(row) && (
-            <OsdsButton
-              color={ODS_THEME_COLOR_INTENT.primary}
-              type={ODS_BUTTON_TYPE.button}
-              size={ODS_BUTTON_SIZE.sm}
-              variant={ODS_BUTTON_VARIANT.stroked}
-              onClick={() => navigate(`./monthly?instanceId=${row.instanceId}`)}
-            >
-              {t('cpbc_hourly_instance_pass_to_monthly')}
-            </OsdsButton>
-          )}
-        </div>
-      ),
-      label: '',
-    },
-  ].filter(Boolean);
 
   const isPending = allInstancesPending || allImagesPending || catalogPending;
