+++ packages/manager/apps/pci-billing/src/components/consumption/VolumeDetailPopover.component.tsx
@@ -3,7 +3,7 @@ import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 import { ODS_TEXT_LEVEL, ODS_TEXT_SIZE } from '@ovhcloud/ods-components';
 import { OsdsText } from '@ovhcloud/ods-components/react';
 import { useTranslation } from 'react-i18next';
-import { TMappedVolume } from './VolumeList.component';
+import { TMappedVolume } from './useVolumeListColumns';
 
 export default function VolumeDetailPopover({
   row,
